backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
google-generativeai>=0.8.3
google-adk>=1.0.0
pydantic>=2.9.2
python-dotenv>=1.0.1
httpx>=0.27.2
